Section-5/Multi-Agent-3/requirements.txt
crewai
crewai-tools
boto3
python-dotenv
google-generativeai